-made save/restore state work; screen rotate no longer resets everything -factored path following code in MainActivity -MainActivity made labelAliasChanged, onCodeInPathSelected, codeSelected more defensive -removed redundant parameter in codeSelected
now when having long click menu of a code in a field open, if you rotate the screen logcat shows this:
02-28 23:30:00.661: E/WindowManager(22748): android.view.WindowLeaked: Activity com.example.kore.ui.MainActivity has leaked window android.widget.PopupWindow$PopupViewContainer@41ab4c80 that was originally added here

while testing I thought I saw the tree resetting to an old version when rotating the screen. I put a breakpoint on onSaveInstanceState, and it wasn't getting called. can't reproduce anymore

also saw this in logcat while breakpointing onSaveInstanceState and keeping execution suspended for a while:
03-01 00:31:07.752: E/ActivityThread(24668): java.lang.RuntimeException: Performing pause of activity that is not resumed: {com.example.kore/com.example.kore.ui.MainActivity}